requirements.txt
streamlit==1.28.0
streamlit-webrtc
av
opencv-python-headless
tensorflow
pillow
numpy
scikit-learn
matplotlib
pyngrok
requests
